Create loop for dots, add event listener, add header, style body
Changes:

index.html
added header

scripts.js
Use a for loop to create the dots that will change color. Added event
listener to each one of the dots and a function that changes their color

styles.css

Added style for the body, made the container a grid, and added style to
the dot class.